requirements.txt
django==1.8.6
django-grappelli==2.7.2
celery
django-celery
requests
django-model-utils